+++ package.json
@@ -7,13 +7,17 @@
   "repository": "git://github.com/superwolff/metalsmith-in-place.git",
   "version": "2.0.0-beta.1",
   "scripts": {
+    "build": "babel -d build/ lib/",
+    "prepublish": "npm run build",
     "debug": "DEBUG=metalsmith-in-place mocha --reporter spec",
-    "test:eslint": "eslint '{src,lib}/**/*.js'",
+    "test:lint": "eslint '{lib,test}/**/*.js'",
     "test:test": "mocha --reporter spec",
-    "test": "npm-run-all test:*"
+    "test": "npm-run-all build test:*"
   },
   "devDependencies": {
     "assert-dir-equal": "^1.0.1",
+    "babel-cli": "^6.16.0",
+    "babel-preset-es2015": "^6.16.0",
     "eslint": "^3.0.1",
     "eslint-config-airbnb-base": "^4.0.0",
     "eslint-plugin-import": "^1.10.3",
